requirements.txt
typer
requests
fastapi
uvicorn
passlib[bcrypt]
bcrypt==4.0.1
python-jose